requirements.txt
Flask>=3.0,<4.0
openai>=1.30,<2.0
python-dotenv>=1.0,<2.0
slack-bolt>=1.23,<2.0